Move JsRuntimeScheduler to JsRuntime
