+++ Plugins/XMLHttpRequest/Source/XMLHttpRequest.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Babylon/JsRuntime.h>
+#include <Babylon/JsRuntimeScheduler.h>
 
 #include <napi/napi.h>
 #include <arcana/threading/task.h>
@@ -121,25 +122,7 @@ namespace Babylon
         arcana::task<void, std::exception_ptr> SendAsyncImpl(); // TODO: Eliminate this function once the UWP file access bug is fixed.
         void SetReadyState(ReadyState readyState);
 
-        class RuntimeScheduler
-        {
-        public:
-            explicit RuntimeScheduler(JsRuntime& runtime)
-                : m_runtime{runtime}
-            {
-            }
-
-            template<typename CallableT>
-            void operator()(CallableT&& callable) const
-            {
-                m_runtime.Dispatch([callable{std::forward<CallableT>(callable)}](Napi::Env){
-                    callable();
-                });
-            }
-
-        private:
-            JsRuntime& m_runtime;
-        } m_runtimeScheduler;
+        JsRuntimeScheduler m_runtimeScheduler;
 
         const std::string m_rootUrl{};
